Refactor `derive_pass_read_limit` to function.cpp
